+++ src/CoolingMap.cpp
@@ -38,11 +38,11 @@ void CoolingMap::Output() {
   for (int m = 0; m < 3; ++m) {
     out.open(mName + "_" + std::to_string(m) + ".dat");
     for (int t = 0; t < mTemperatures.size(); ++t) {
-      FLOAT temp = mTemperatures[t];
+      float temp = mTemperatures[t];
       for (int d = 0; d < mDensities.size(); ++d) {
-        FLOAT dens = mDensities[d];
-        FLOAT kappa = mOpacity->GetKappa(dens, temp) * MOD_ARRAY[m];
-        FLOAT dudt = CalculateDUDT(dens, temp, kappa);
+        float dens = mDensities[d];
+        float kappa = mOpacity->GetKappa(dens, temp) * MOD_ARRAY[m];
+        float dudt = CalculateDUDT(dens, temp, kappa);
         out << dens << "\t" << temp << "\t" << log10(dudt) << "\n";
       }
     }
@@ -53,12 +53,12 @@ void CoolingMap::Output() {
   for (int m = 0; m < 3; ++m) {
     out.open(mName + "_" + std::to_string(m) + "_contour.dat");
     for (int t = 0; t < mTemperatures.size(); ++t) {
-      FLOAT temp = mTemperatures[t];
-      FLOAT cur_tau = 0.0f;
+      float temp = mTemperatures[t];
+      float cur_tau = 0.0f;
       for (int d = 0; d < mDensities.size(); ++d) {
-        FLOAT dens = mDensities[d];
-        FLOAT kappa = mOpacity->GetKappa(dens, temp) * MOD_ARRAY[m];
-        FLOAT tau = kappa * dens * AU_TO_CM;
+        float dens = mDensities[d];
+        float kappa = mOpacity->GetKappa(dens, temp) * MOD_ARRAY[m];
+        float tau = kappa * dens * AU_TO_CM;
         if (cur_tau < 1.0f && tau > 1.0f) {
           out << 0.5f * (dens + mDensities[d - 1]) << "\t"
               << 0.5f * (temp + mTemperatures[t - 1]) << "\t"
@@ -75,19 +75,19 @@ void CoolingMap::Output() {
 void CoolingMap::FillVectors() {
   float dens_step = fabs(mDensMax - mDensMin) / (float)mDensBins;
   for (int i = 0; i < mDensBins; ++i) {
-    mDensities.push_back(powf(10.0, mDensMin + i * dens_step));
+    mDensities.push_back(pow(10.0, mDensMin + i * dens_step));
   }
 
   float temp_step = fabs(mTempMax - mTempMin) / (float)mTempBins;
   for (int i = 0; i < mTempBins; ++i) {
-    mTemperatures.push_back(powf(10.0, mTempMin + i * temp_step));
+    mTemperatures.push_back(pow(10.0, mTempMin + i * temp_step));
   }
 }
 
-FLOAT CoolingMap::CalculateDUDT(const FLOAT dens, const FLOAT temp,
-                                const FLOAT kappa) {
-  const FLOAT sigma = dens * AU_TO_CM;
-  const FLOAT numer = 4.0f * SB * powf(temp, 4.0f);
-  const FLOAT denom = (sigma * sigma * kappa) + (1.0f / kappa);
+float CoolingMap::CalculateDUDT(const float dens, const float temp,
+                                const float kappa) {
+  const float sigma = dens * AU_TO_CM;
+  const float numer = 4.0f * SB * pow(temp, 4.0f);
+  const float denom = (sigma * sigma * kappa) + (1.0f / kappa);
   return numer / denom;
 }
